Create bubblesort.py
Nevertheless, this sorting algorithm is commonly used where the value of n is not too large and programming effort is to be kept to a minimum. The bubble sort has the additional virtue that it requires almost no space in addition to that used to contain the input vector. 

Bubble Sort: Where we know the data is very nearly sorted. Say only two elements are out of place. Then in one pass, BS will finish it and in the second pass, it will see everything is sorted and then exit. Only takes 2 passes of the array. 

Performance: Although popular, bubble sort is nearly universally derided for its poor performance on random data. This is evident as shown below. Some books laud bubble sort because it runs in O(n) time on sorted data and works well on ``nearly sorted'' data.
  
referred: https://users.cs.duke.edu/~ola/bubble/bubble.html
Bubble sort should be avoided in the case of large collections. It will not be efficient in the case of a reverse-ordered collection