Create insertion.py
Insertion sort is a sorting algorithm that builds a final sorted array (sometimes called a list) one element at a time. While sorting is a simple concept, it is a basic principle used in complex computer programs such as file search, data compression, and path finding. Running time is an important thing to consider when selecting a sorting algorithm since efficiency is often thought of in terms of speed. Insertion sort has an average and worst-case running time of O(n^2), so in most cases, a faster algorithm is more desirable. 
•	Efficient for (quite) small data sets, much like other quadratic sorting algorithms
•	More efficient in practice than most other simple quadratic (i.e., O(n2)) algorithms such as selection sort or bubble sort
•	Adaptive, i.e., efficient for data sets that are already substantially sorted: the time complexity is O(nk) when each element in the input is no more than k places away from its sorted position
•	Stable; i.e., does not change the relative order of elements with equal keys
•	In-place; i.e., only requires a constant amount O(1) of additional memory space
•	Online; i.e., can sort a list as it receives it